chore: split out styled components
