+++ packages/native-ui/src/components/custom/Spinner/styled-components/Circle.ts
@@ -1,0 +1,13 @@
+import { styled } from '@gluestack-style/react';
+import Animated from 'react-native-reanimated';
+import { Circle } from 'react-native-svg';
+
+const AnimatedCircle = Animated.createAnimatedComponent(Circle);
+
+const SpinnerCircle = styled(
+  AnimatedCircle,
+  {},
+  { componentName: 'SpinnerCircle', ancestorStyle: ['_circle'] }
+);
+
+export default SpinnerCircle;
