+++ packages/native-ui/src/components/custom/Spinner/styled-components/Svg.ts
@@ -1,0 +1,17 @@
+import { styled } from '@gluestack-style/react';
+import Animated from 'react-native-reanimated';
+import { Svg } from 'react-native-svg';
+
+const AnimatedSvg = Animated.createAnimatedComponent(Svg);
+
+const SpinnerSVG = styled(
+  AnimatedSvg,
+  {},
+  {
+    componentName: 'SpinnerSvg',
+    ancestorStyle: ['_svg'],
+    descendantStyle: ['_circle'],
+  }
+);
+
+export default SpinnerSVG;
