+++ packages/native-ui/src/components/custom/Spinner/styled-components/index.ts
@@ -1,0 +1,3 @@
+export { default as StyledSpinner } from './Root';
+export { default as StyledCircle } from './Circle';
+export { default as StyledSvg } from './Svg';
